% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.corr.R
\name{auto.corr}
\alias{auto.corr}
\title{Calculating autocorrelation}
\usage{
auto.corr(x, model)
}
\arguments{
\item{x}{vector of sample means}

\item{model}{the model being evaluated}
}
\value{
the autocorrelation
}
\description{
Estimates the autocorrelation in the data.
This function is used by other functions and will generally not be used directly by users.
}
\details{
This function calculates the autocorrelation in a vector of sample means
defined as the correlation of the first n-1 observations with the last n-1. The
autocorrelation is calculated directly on the sample means if the evaluated model is stasis.
If a different model is evaluated (random walk or directional trend), the data is
detrended prior to the calculation of autocorrelation.
}
