% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.corr.test.BM.R
\name{auto.corr.test.BM}
\alias{auto.corr.test.BM}
\title{Applying the autocorrelation test to the Random walk model}
\usage{
auto.corr.test.BM(y, vstep, nrep = 1000, conf = 0.95, plot = TRUE,
  save.replicates = TRUE)
}
\arguments{
\item{y}{a paleoTS object}

\item{vstep}{the variance of the step distribution estmated from the observed data.}

\item{nrep}{number of iterations in the parametric boostrap (number of simulated time series); default is 1000.}

\item{conf}{confidence level for judging whether a model is an adequate statistical description of the data.
Number must be between 0 and 1. A higher number means less strict judgment of whether a model is adequate; default
is 0.95. Tests are two-tailed (except for the net evolution test), which means a model is judged adequate if the observed test statistic is within the 2.5
percent of the extreeme values of the calculated test statistics on the simulated data given the default confidence
value of 0.95.}

\item{plot}{logical; if TRUE, the value of the test statistic calculated based on the observed fossil
time series is plotted on the distribution of test statistics calculated on the simulated time series;
default is TRUE.}

\item{save.replicates}{logical; if TRUE, the values of the test statistic calculated on the simulated time
series is saved and can be accessed later for plotting purposes; default is TRUE.}
}
\value{
First part of the output summarizes the number of iterations in the parametric boostrap and the
confidence level for judging whether a model is an adequate statistical description of the data. The last
part of the output is:

\item{estimate}{The calculated test statistic on the observed data.}
 \item{min.sim}{The smallest test statistic calculated on the simulated data.}
 \item{max.sim}{The largest test statistic calculated on the simulated data.}
 \item{p-value}{Not a real p-value, but is calculated as the fraction of simulated test statistics
 that is larger (or smaller) than the calculated test statistic on the observed data divided by 0.5.
 A value of 1 means 50 percent of the test statistics on the simulated data are largen and smaller
 than the calculated statistic on the observed data. A value of 0.10 means 90 percent of the test
 statistics on the simulated data are larger or smaller than the test statistic on the observed time
 series.}
 \item{result}{Whether the model PASSED or FAILED the adequasy test. The outcome depends on the
 confidence level.}
}
\description{
Investigates if the Random walk model is an adequate statistical description of an evolutionary
time series by applying the autocorrelation test.
}
\details{
This function calculates the autocorrelation in a vector of sample means
defined as the correlation of the first n-1 observations with the last n-1. The
aurocorrelation is calculated directly on the sample means if the evaluated model is stasis.
If a different model is evaluated (random walk or directional trend), the data is
detrended prior to the calculation of autocorrelation.
}
\examples{
## generate a paleoTS objects by simulating a directional trend
x <- sim.GRW(ns=40, ms=0, vs=0.1)

## estimate the variance of the step distribution
vstep <- mle.GRW(x)[2]

## investigate if the time series pass the adequasy test
auto.corr.test.BM(x,vstep)

}
\references{
Voje, K.L., Starrfelt, J., and Liow, L.H. Model adequacy and microevolutionary explanations for stasis in the fossil record. \emph{The American Naturalist}. In press.
}
\seealso{
\code{\link{fit3adequasy.BM}}, \code{\link{auto.corr.test.DT}}, \code{\link{auto.corr.test.stasis}}
}
\author{
Kjetil L. Voje
}
