% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.in.morphospace.R
\name{dist.in.morphospace}
\alias{dist.in.morphospace}
\title{Calculating distances in morphospace.}
\usage{
dist.in.morphospace(y, correct = TRUE, iter = 10000)
}
\arguments{
\item{y}{a paleoTS object}

\item{correct}{logical, if TRUE, distances between population means are corrected for sampling variance.}

\item{iter}{number of iterations in the parametric bootstrap; default is 1000.}
}
\value{
distances in morphospace between sample means in time series.
}
\description{
This function calculates the distances in morphospace between sample means in a time series.
This function is used by other functions and will generally not be used directly by users.
}
\references{
Voje, K.L. 2016. Tempo does not correlate with mode in the fossil record. \emph{Evoluton} 70:2678–2689.
}
\author{
Kjetil L. Voje
}
