% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.OU.R
\name{est.OU}
\alias{est.OU}
\title{Predicting trait values for a given OU model}
\usage{
est.OU(x, w, tt)
}
\arguments{
\item{x}{a paleoTS object}

\item{w}{an object of class paleoTSfit}

\item{tt}{vector of sample ages, increases from oldest to youngest}
}
\value{
Predicted trait values for an OU model with given parameters.
}
\description{
This function calculates the predicted trait values for a given OU model.
This function is used by other functions and will generally not be used directly by users.
}
\author{
Kjetil L. Voje
}
