% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit3adequacy.accel.R
\name{fit3adequacy.accel}
\alias{fit3adequacy.accel}
\title{Applying 3 adequacy tests to the accelerated evolution model}
\usage{
fit3adequacy.accel(y, vstep = NULL, r = NULL, nrep = 1000,
  conf = 0.95, plot = TRUE)
}
\arguments{
\item{y}{a paleoTS object}

\item{vstep}{the variance of the step distribution estimated from the observed data.}

\item{r}{parameter describing the increasing rate change through time. r is restricted to values larger than zero 
(the model reduces to the BM model when r = 0).}

\item{nrep}{number of iterations in the parametric bootstrap (number of simulated time series); default is 1000.}

\item{conf}{confidence level for judging whether a model is an adequate statistical description of the data.
Number must be between 0 and 1. A higher number means less strict judgment of whether a model is adequate; default
is 0.95. Tests are two-tailed, which means a model is judged adequate if the observed test statistic is within the 2.5
percent of the extreme values of the calculated test statistics on the simulated data given the default confidence
value of 0.95.}

\item{plot}{logical; if TRUE, the value of the test statistic calculated based on the observed fossil
time series is plotted on the distribution of test statistics calculated on the simulated time series;
default is TRUE.}
}
\value{
First part of the output summarizes the number of iterations in the parametric bootstrap and the
confidence level for judging whether a model is an adequate statistical description of the data. The last
part of the output is a data frame with the adequacy tests as columns and the following rows:

\item{estimate}{The calculated test statistic on the observed data.}
 \item{min.sim}{The smallest test statistic calculated on the simulated data.}
 \item{max.sim}{The largest test statistic calculated on the simulated data.}
 \item{p-value}{Not a real p-value, but is calculated as the fraction of simulated test statistics
 that is larger (or smaller) than the calculated test statistic on the observed data divided by 0.5.
 A value of 1 means 50 percent of the test statistics on the simulated data are larger and smaller
 than the calculated statistic on the observed data. A value of 0.10 means 90 percent of the test
 statistics on the simulated data are larger or smaller than the test statistic on the observed time
 series.}
 \item{result}{Whether the model PASSED or FAILED the adequasy test. The outcome depends on the
 confidence level.}
}
\description{
Investigating if the accelerated evolution model is an adequate statistical description of an evolutionary
time series by applying the following tests (1) autocorrelation (2) runs test, and (3) test of accelerated evolution.
}
\details{
A wrapper function for investigating the adequacy of the accelerated rate evolution model by
applying all three adequacy tests at the same time.
}
\examples{
## generate a paleoTS objects by simulating early burst
x <- sim.accel_decel(ns=40, r=-1, vs=0.1)

## Investigate if the time series pass all thee adequacy tests
fit3adequacy.accel(x)

}
\references{
Voje, K.L. 2018. Assessing adequacy of models of phyletic evolution in the fossil record. \emph{Methods in Ecology and Evoluton}. (in press).

Voje, K.L., Starrfelt, J., and Liow, L.H. 2018. Model adequacy and microevolutionary explanations for stasis in the fossil record. \emph{The American Naturalist}. 191:509-523.
}
\seealso{
\code{\link{fit3adequasy.trend}}, \code{\link{fit4adequasy.stasis}}
}
\author{
Kjetil L. Voje
}
