% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.change.test.R
\name{net.change.test}
\alias{net.change.test}
\title{Calculating net evolution}
\usage{
net.change.test(x, model)
}
\arguments{
\item{x}{vector of sample means}

\item{model}{the model being evaluated}
}
\value{
net evolution
}
\description{
Calculates the net evolution in the data, i.e. the absolute difference
in trait value between the first and last sample mean in the time series.
This function is used by other functions and will generally not be used directly by users.
}
\details{
This function calculates the net evolution in a vector of sample means
defined as the the absolute difference in trait value between the first and last
sample mean in the time series.
}
