% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.distributions.R
\name{plotting.distributions}
\alias{plotting.distributions}
\title{Plotting the observed test statistic on the distribution of the test statistic calculated on simulated data}
\usage{
plotting.distributions(sim.data, obs, test, xlab, main)
}
\arguments{
\item{sim.data}{vector containing test statistics calculated from the simulated data}

\item{obs}{the test statistic calculated from the observed data}

\item{test}{the test that has been applied (runs.test, slope.test, auto.corr, net.change.test)}

\item{xlab}{label for x-axis}

\item{main}{plot label}
}
\value{
Histogram of test statistics calculated from simulated data. The test statistic calculated from
the observed data is indicated by a vertical broken line.
}
\description{
This function is used by other functions and will generally not be used directly by users. Can be used to plot
test statistics on simulated data generated by other functions (e.g. \code{\link{slope.test.stasis}},
 \code{\link{auto.corr.test.trend}}, \code{\link{runs.test.RW}}).
}
\details{
This function is used by other functions and will generally not be used directly by users. Can be used to plot
test statistics on simulated data generated by other functions (e.g. \code{\link{slope.test.stasis}},
 \code{\link{auto.corr.test.trend}}, \code{\link{runs.test.RW}}).
}
\examples{
## generate a paleoTS objects by simulating a directional trend
x <- sim.GRW(ns=40, ms=0.5, vs=0.1)

## run the test
for.plotting<-auto.corr.test.trend(x,mstep,vstep)

## plot
plotting.distributions(for.plotting$replicates, for.plotting$summary[1], test="auto.corr")

}
\seealso{
\code{\link{fit3adequacy.trend}}, \code{\link{auto.corr.test.RW}}, \code{\link{auto.corr.test.stasis}}
}
\author{
Kjetil L. Voje
}
