% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.test.stasis.R
\name{runs.test.stasis}
\alias{runs.test.stasis}
\title{Applying the runs test to the stasis model}
\usage{
runs.test.stasis(y, theta, omega, nrep = 1000, conf = 0.95, plot = TRUE,
  save.replicates = TRUE)
}
\arguments{
\item{y}{a paleoTS object}

\item{theta}{evolutionary optimum estmated from the observed data.}

\item{omega}{evolutonary variance estmated from the observed data.}

\item{nrep}{number of iterations in the parametric boostrap (number of simulated time series); default is 1000.}

\item{conf}{confidence level for judging whether a model is an adequate statistical description of the data.
Number must be between 0 and 1. A higher number means less strict judgment of whether a model is adequate; default
is 0.95. Tests are two-tailed (except for the net evolution test), which means a model is judged adequate if the observed test statistic is within the 2.5
percent of the extreeme values of the calculated test statistics on the simulated data given the default confidence
value of 0.95.}

\item{plot}{logical; if TRUE, the value of the test statistic calculated based on the observed fossil
time series is plotted on the distribution of test statistics calculated on the simulated time series;
default is TRUE.}

\item{save.replicates}{logical; if TRUE, the values of the test statistic calculated on the simulated time
series is saved and can be accessed later for plotting purposes; default is TRUE.}
}
\value{
First part of the output summarizes the number of iterations in the parametric boostrap and the
confidence level for judging whether a model is an adequate statistical description of the data. The last
part of the output is:

\item{estimate}{The calculated test statistic on the observed data.}
 \item{min.sim}{The smallest test statistic calculated on the simulated data.}
 \item{max.sim}{The largest test statistic calculated on the simulated data.}
 \item{p-value}{Not a real p-value, but is calculated as the fraction of simulated test statistics
 that is larger (or smaller) than the calculated test statistic on the observed data divided by 0.5.
 A value of 1 means 50 percent of the test statistics on the simulated data are largen and smaller
 than the calculated statistic on the observed data. A value of 0.10 means 90 percent of the test
 statistics on the simulated data are larger or smaller than the test statistic on the observed time
 series.}
 \item{result}{Whether the model PASSED or FAILED the adequasy test. The outcome depends on the
 confidence level.}
}
\description{
Investigates if the stasis model is an adequate statistical description of an evolutionary
time series by applying the runs test.
}
\details{
This function applies a runs test in order to investigate if the stasis model can be judged an
adequate statistical description of the data. In the stasis model, there should be no tendency for a trait to successively deviate
from the optimum in the same direction and the runs test is applied to the sign of the residuals (i.e. θ – trait value)
to identify series that have non-random patterns in the sign of deviations. For a time series of length n,
the number of runs (one run is a sequence of consecutive numbers with same sign), is approximately normal
with mean μ=(2(n_+ n_-))/n+1 and variance (μ-1)(μ-2)/(n-1), where n+ and n- are the number of residuals
above and below the optimum respectively. The mean and variance are used to calculate the standard/Z-score
implemented as the test statistic.
}
\examples{
## generate a paleoTS objects by simulating a stasis time series
x <- sim.Stasis(ns = 40, theta = 0, omega = 0.1)

## estimate the evolutionary optimum
theta <- mle.Stasis(x)[1]

## estimate the evolutionary variance
omega <- mle.Stasis(x)[2]

## investigate if the time series pass the adequasy test
runs.test.stasis(x,theta,omega)

}
\references{
Voje, K.L., Starrfelt, J., and Liow, L.H. Model adequacy and microevolutionary explanations for stasis in the fossil record. \emph{The American Naturalist}. In press.
}
\seealso{
\code{\link{fit3adequasy.RW}}, \code{\link{fit3adequasy.trend}}
}
\author{
Kjetil L. Voje
}
