% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.accel_decel.R
\name{sim.accel_decel}
\alias{sim.accel_decel}
\title{Simulate time-series where the rate of evolution increase or decrease with time}
\usage{
sim.accel_decel(ns = 20, r = -5, vs = 0.1, vp = 0.01,
  nn = rep(20, ns), tt = 0:(ns - 1)/ns)
}
\arguments{
\item{ns}{number of samples in time-series}

\item{r}{parameter describing the decreasing/increasing rate of change through time.}

\item{vs}{variance of the step distribution.}

\item{vp}{within-population trait variance.}

\item{nn}{vector of the number of individuals in each sample}

\item{tt}{vector of sample ages, increases from oldest to youngest}
}
\value{
First part of the output summarizes the number of iterations in the parametric bootstrap and the
confidence level for judging whether a model is an adequate statistical description of the data. The last
part of the output is a data frame with the adequacy tests as columns and the following rows:

A paleoTS object.
}
\description{
Simulate the evolution of a trait according to a random walk where where the rate of evolution increase or decrease with time.
}
\details{
The function simulate the evolution of a trait where the rate of evolution increase or decrease with time:
The trait evolves accoridng to a random walk where the rate of evolution show either an exponential
increase or decrease with time. See reference below for details on the model.
}
\examples{
## generate a paleoTS objects by simulating a trait that shows an increasing rate of evolution with time. 
x <- sim.accel_decel(ns=20, r=-1, vs=0.1)

## Investigate if the time series pass all thee adequacy tests
fit3adequacy.accel(x)

}
\references{
Harmon, L. J., J. B. Losos, T. J. Davies, R. G. Gillespie, J. L. Gittleman, W. B. Jennings, K. H. Kozak, et al. 2010. Early bursts of body size and shape evolution are rare in comparative data. \emph{Evolution} 64:2385–2396.
}
\author{
Kjetil L. Voje
}
