% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.EB.R
\name{sim.EB}
\alias{sim.EB}
\title{Simulate early burst time-series}
\usage{
sim.EB(ns = 20, alpha = -1, vs = 0.1, vp = 0.01, nn = rep(20,
  ns), tt = 0:(ns - 1))
}
\arguments{
\item{ns}{number of samples in time-series}

\item{alpha}{parameter describing the decreasing rate of change through time.}

\item{vs}{variance of the step distribution.}

\item{vp}{within-population trait variance.}

\item{nn}{vector of the number of individuals in each sample}

\item{tt}{vector of sample ages, increases from oldest to youngest}
}
\value{
First part of the output summarizes the number of iterations in the parametric bootstrap and the
confidence level for judging whether a model is an adequate statistical description of the data. The last
part of the output is a data frame with the adequacy tests as columns and the following rows:

A paleoTS object.
}
\description{
Simulate the evolution of a trait according to the early burst model.
}
\details{
The functin simulate the evolution of a trait according to the early burst model: A model with 
initial higher rates of evolution (early burst) followed by an exponential drop in the rate of evolution. 
See reference below for details on the model.
}
\examples{
## generate a paleoTS objects by simulating early burst
x <- sim.EB(ns=20, alpha=-1, vs=0.1)

## Investigate if the time series pass all thee adequacy tests
fit3adequacy.EB(x)

}
\references{
Harmon, L. J., J. B. Losos, T. J. Davies, R. G. Gillespie, J. L. Gittleman, W. B. Jennings, K. H. Kozak, et al. 2010. Early bursts of body size and shape evolution are rare in comparative data. \emph{Evolution} 64:2385–2396.
}
\author{
Kjetil L. Voje
}
