% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.test.R
\name{slope.test}
\alias{slope.test}
\title{Estimating the linear time-dependency of the trait variance}
\usage{
slope.test(x, model, tt, theta = NULL, int = NULL, mstep = NULL)
}
\arguments{
\item{x}{vector of sample means}

\item{model}{the model being evaluated}

\item{tt}{vector of sample ages}

\item{theta}{if the tested model is stasis, theta is the estimated theta from the data.}
}
\value{
least-square slope estimate
}
\description{
This function estimates the linear dependency of the trait variance on time.
This function is used by other functions and will generally not be used directly by users.
}
\details{
If model = stasis: Estimates the slope of the least square regression of the size of deviations
(their absolute value) from the optimal phenotype as a function of time. If model = random walk or directional
trend: Estimates the slope of the least square regression of the size of the detrended data as a function of time.
This function is used by other functions and will generally not be used directly by users.
}
